bringup: the soc level done, driver level pending
I did my best to bring this as close as possible to the BSP36 u-boot configuration.
It is pretty straightforward, soc/arch level stuff are in board_f.c, that all initializes now
driver level stuff is all in board_r.c, I am still working on that but progress so far has been
pretty good. We are almost at the point where we can have an interractive console, have to
tweak and hack together some more subsystems but also possibly some MMU related things as currently
I am getting this error when I try to initialize a peripheral like the network interface or the MMC controller:

Net:   PTE 0000000000000000 (0) for addr=d0000000 should be a table
PTE 0000000000000000 (0) for addr=d0000000 should be a table