bringup: we can reach the interactive console now
This brings 2 things:

1. Build in the initial board file object s32gxxxardb.o that
will be used as a basis to pull more things from the s32g274ar2sbc2
machine from the Microsis patches. Those are *extremely* hard to port
as-is and we'll have a better luck including little by little what we
need from there. In principle anything GMAC related so that we can finally
work with the SJA1110 ethernet switch.

2. early_mmu_init() in arch_cpu_init() during early boot is buggy. It leaves
the TLB address to a location pointing into the TEXT area of the u-boot binary
itself. This was *very* annoying to debug since it leads to really subtle issues
during page address translation. The only point of that function is to enable caches early,
we'll drop it since we enable I and D caches not too much later in the boot process anyway.