Bring up GMAC0, add some commands to drivers/misc/sja1105.c
sja command has 3 new commands now.

The first one is the "init_ports" command, which attempts to configure
the port 3 of the ethernet switch as SGMII and port 1 as 100BASE-TX
The port 1 at least seems to respond to the ethernet interface to which it is
connected's link status. If the other end is down the link doesn't come up.
If the tw0030's interface is brought back up, the link UP status can be detected
from the N4's ethernet switch..

sja also has 2 new commands called "spi" and "sgmii".
I use these to easily read/write these registers without needing to rebuild.
The advantage of these commands is that you can use the same register offsets
that you read in the datasheet.

An example of "sja sgmii" command is:
=> sja sgmii 0 3 1f80 40 2:0
RD[0x1f8040] 0xbe5

The first 0 means "read", if you write non zero it means "write"
The 3 is the SGMII port number, 1f80 is the base address that is to be written
in the indirection SGMII base address register, which is the SGMII register's address
shifted by 8 bits to the right. For example if you want to access 0x1f8040
you will write 0x1f80 in the base register and 0x40 in the offset.
The last argument is the usual bus:chipselect SPI address.

If it's a write command you write "1" instead of "0" for the first argument and the
value to be written comes before the SPI address.
Eg, writing 0xff to the same register:
=> sja sgmii 1 3 1f80 40 0xff 2:0

The "spi" subcommand is similar, it just doesn't need a port number, and the base
address is just the full base address you'd see in the datasheet for the given subsystem.

Eg if you want to access the Auxiliary Control Unit (ACU) you use base address 0x1c4400

and then the offset field is the register as it is in the datasheet.

For eg reading PIN_CFG_MII2_ID which is reg 18h:

=> sja spi 0 1c4400 18 2:0
RD[0x1c4418] 0x1110111

Which is the default value as according to the datasheet.