Disable SJA1110 microcontroller
The Linux driver seems to not use the switch's MCU at all.
Disable it by disabling the watchdog timer and MCU clock lines.

I re-introduced the load_config step upon driver probe since that
was just means as a debug, will keep the "load_config" command
but it is uneccesary. It should be enough to simply do:

sja probe 2:0
sja init_ports 2:0

to bring the SGMII and 100BASE-TX ports up